ipykernel-kernel-service/slave/requirements.txt
pandas
numpy
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
dnspython==2.5.0
docker==7.0.0
executing==2.0.1
fastapi==0.109.0
h11==0.14.0
httptools==0.6.1
idna==3.6
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.35.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.27.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
scikit-learn
matplotlib